Upgrade drone to mongo 3.0
This repo will actually connect to 2.4 and 3.0 depending on the data
type, but I thought it would be better to have the tests run on 3.0
since that's the new. We could make it run it against both, but I
thought that was more effort than it was worth. Let me know if you
disagree.