Re-enable using projections in mongo queries.
Bug was found, if data.name.first exists in the projection, it blocks data.name, if it's not an object.
Instead, we'll flag this on the table level in ark-config, so we can get speedup on some tables, without affecting them all.

Revert "Revert "Only request the desired fields from the mongo db.""

This reverts commit 311e5d11b17937055ea3d13d05bde1a8d5c553ac.